extract downloadFile into separate file to DRY up
