+++ scripts/fetch-binaries.js
@@ -1,9 +1,8 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 const fs = require('fs');
 const path = require('path');
-const crypto = require('crypto');
 const tsNode = require('ts-node');
-const childProcess = require('child_process');
+const downloadFile = require('./download');
 
 tsNode.register();
 
@@ -45,33 +44,6 @@ const holochainBinaryFilename = `holochain-v${
 const lairBinaryFilename = `lair-keystore-v${kangarooConfig.bins.lair.version}-${binariesAppendix}${
   process.platform === 'win32' ? '.exe' : ''
 }`;
-
-function downloadFile(url, targetPath, expectedSha256Hex, chmod = false) {
-  console.log('Downloading from ', url);
-  childProcess.exec(`curl -f -L --output ${targetPath} ${url}`, (error, stdout, stderr) => {
-    console.log(stdout);
-    console.log(stderr);
-    if (error !== null) {
-      console.log('exec error: ' + error);
-      throw new Error('Failed to fetch resource.');
-    } else {
-      const fileBytes = fs.readFileSync(targetPath);
-      const hasher = crypto.createHash('sha256');
-      hasher.update(fileBytes);
-      const sha256Hex = hasher.digest('hex');
-      if (sha256Hex !== expectedSha256Hex)
-        throw new Error(
-          `sha256 does not match the expected sha256. Got ${sha256Hex} but expected ${expectedSha256Hex}`
-        );
-
-      console.log('Download successful. sha256 of file (hex): ', sha256Hex);
-      if (chmod) {
-        fs.chmodSync(targetPath, 511);
-        console.log('Gave executable permission to file.');
-      }
-    }
-  });
-}
 
 function downloadHolochainBinary() {
   const holochainBinaryRemoteFilename = `holochain-v${kangarooConfig.bins.holochain.version}-${targetEnding}`;
